% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example.A}
\alias{Example.A}
\title{Example.A dataset}
\format{
A data frame with 2800 rows and 17 variables:
\describe{
  \item{ID}{Numeric, case ID number.}
  \item{Region}{Character, Grouping variable for Groups: European region ("North West", "Nordic", "Continental", "Southern", "Central East", "North East" and "South East").}
  \item{Org Size}{Numeric, control variable - organization size.}
  \item{Tenure}{Numeric, control variable - tenure in years.}
  \item{R45a}{Numeric, Work-kife conflict item R45a from 1 (always) to 5 (never)}
  \item{R45b}{Numeric, Work-kife conflict item R45b from 1 (always) to 5 (never)}
  \item{R45c}{Numeric, Work-kife conflict item R45c from 1 (always) to 5 (never)}
  \item{R45d}{Numeric, Work-kife conflict item R45d from 1 (always) to 5 (never)}
  \item{R45e}{Numeric, Work-kife conflict item R45e from 1 (always) to 5 (never)}
  \item{R90a}{Numeric, Engagement item R90a from 1 (always) to 5 (never)}
  \item{R90b}{Numeric, Engagement item R90b from 1 (always) to 5 (never)}
  \item{R90c}{Numeric, Engagement item R90c from 1 (always) to 5 (never)}
  \item{R87a}{Numeric, Psychological wellbeing item R87a from 1 (all of the time) to 6 (at no time)}
  \item{R87b}{Numeric, Psychological wellbeing item R87b from 1 (all of the time) to 6 (at no time)}
  \item{R87c}{Numeric, Psychological wellbeing item R87c from 1 (all of the time) to 6 (at no time)}
  \item{R87d}{Numeric, Psychological wellbeing item R87d from 1 (all of the time) to 6 (at no time)}
  \item{R87e}{Numeric, Psychological wellbeing item R87e from 1 (all of the time) to 6 (at no time)}
}
}
\source{
Simulated dataset (400 observations for each region) based on the 6th European Working Conditions Survey (EWCS 2015; Eurofound, 2024) -- Examined the life and working conditions of 43,850 respondents in 35 European countries, which were divided into seven regions: North West, Nordic, Continental, Southern, Central East, North East and South East
}
\usage{
Example.A
}
\description{
Simulated dataset for demonstrating Full_MEI(), CompareLoadings(), CompareMeans(), and CompareParameters()
}
\keyword{datasets}
