% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example.B}
\alias{Example.B}
\title{Example.B dataset}
\format{
A data frame with 242 rows and 16 variables:
\describe{
  \item{ID}{Numeric, case ID number.}
  \item{x1_T1}{Numeric, Satisfaction with life scale Item 1 (SWLS1) at T1 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x2_T1}{Numeric, Satisfaction with life scale Item 2 (SWLS1) at T1 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x3_T1}{Numeric, Satisfaction with life scale Item 3 (SWLS1) at T1 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x4_T1}{Numeric, Satisfaction with life scale Item 4 (SWLS1) at T1 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x5_T1}{Numeric, Satisfaction with life scale Item 5 (SWLS1) at T1 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x1_T2}{Numeric, Satisfaction with life scale Item 1 (SWLS1) at T2 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x2_T2}{Numeric, Satisfaction with life scale Item 2 (SWLS1) at T2 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x3_T2}{Numeric, Satisfaction with life scale Item 3 (SWLS1) at T2 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x4_T2}{Numeric, Satisfaction with life scale Item 4 (SWLS1) at T2 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x5_T2}{Numeric, Satisfaction with life scale Item 5 (SWLS1) at T2 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x1_T3}{Numeric, Satisfaction with life scale Item 1 (SWLS1) at T3 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x2_T3}{Numeric, Satisfaction with life scale Item 2 (SWLS1) at T3 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x3_T3}{Numeric, Satisfaction with life scale Item 3 (SWLS1) at T3 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x4_T3}{Numeric, Satisfaction with life scale Item 4 (SWLS1) at T3 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
  \item{x5_T3}{Numeric, Satisfaction with life scale Item 5 (SWLS1) at T3 from 1 (lowest satisfaction) to 6 (greatest satisfaction)}
}
}
\source{
Simulated dataset with 242 observations across three time points based on Study 2 in Wu, C., Chen, L. H., & Tsai, Y. (2009). Longitudinal invariance analysis of the satisfaction with life scale. Personality and Individual Differences, 46: 396-401.
}
\usage{
Example.B
}
\description{
Simulated dataset for demonstrating LGCompareLoadings() and LGCompareMeans() in longitudinal studies
}
\keyword{datasets}
