% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example.C}
\alias{Example.C}
\title{Example.C dataset}
\format{
A data frame with 332 rows and 21 variables:
\describe{
  \item{ID}{Numeric, case ID number.}
  \item{x1_G1}{Numeric, Self-ratings of managerial role item 1  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x2_G1}{Numeric, Self-ratings of managerial role item 2  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x3_G1}{Numeric, Self-ratings of managerial role item 3  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x4_G1}{Numeric, Self-ratings of managerial role item 4  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x5_G1}{Numeric, Self-ratings of managerial role item 5  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x6_G1}{Numeric, Self-ratings of managerial role item 6  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x7_G1}{Numeric, Self-ratings of managerial role item 7  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x8_G1}{Numeric, Self-ratings of managerial role item 8  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x9_G1}{Numeric, Self-ratings of managerial role item 9  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x10_G1}{Numeric, Self-ratings of managerial role item 10  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x1_G2}{Numeric, Supervisor-ratings of managerial role item 1  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x2_G2}{Numeric, Supervisor-ratings of managerial role item 2  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x3_G2}{Numeric, Supervisor-ratings of managerial role item 3  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x4_G2}{Numeric, Supervisor-ratings of managerial role item 4  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x5_G2}{Numeric, Supervisor-ratings of managerial role item 5  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x6_G2}{Numeric, Supervisor-ratings of managerial role item 6  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x7_G2}{Numeric, Supervisor-ratings of managerial role item 7  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x8_G2}{Numeric, Supervisor-ratings of managerial role item 8  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x9_G2}{Numeric, Supervisor-ratings of managerial role item 9  from 1 (not at all effective) to 7 (extremely effective).}
  \item{x10_G2}{Numeric, Supervisor-ratings of managerial role item 10  from 1 (not at all effective) to 7 (extremely effective).}
}
}
\source{
Simulated dataset with 332 dyads of mid-level executives and their supervisors based on Ashford, S. J., & Tsui A. S. (1991). Self-regulation for managerial effectiveness: The role of active feedback seeking. Academy of Management Journal, 34, 251-280.
}
\usage{
Example.C
}
\description{
Simulated dataset for demonstrating LGCompareLoadings() and LGCompareMeans() in congruence studies.
}
\keyword{datasets}
