% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI.R
\name{Full_MEI}
\alias{Full_MEI}
\title{Full Measurement Invariance Test}
\usage{
Full_MEI(model, data.source, Groups, Cluster = "NULL", correct.cfi = TRUE)
}
\arguments{
\item{model}{User-specified CFA model.}

\item{data.source}{A data frame containing the observed variables used in the model.}

\item{Groups}{Grouping variable for cross-group comparisons.}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}

\item{correct.cfi}{If TRUE (default), baseline model will be corrected for calculation of incremental fit indices based on Widaman & Thompson (2003).}
}
\value{
lavaan outputs, model fit for configural, metric and scalar invariance models, summary of fit statistics for each model.
}
\description{
Conduct configural invariance, full metric invariance, and full scalar invariance tests for data from independent groups.
}
\details{
Reference for correct.cfi: Widaman, K. F., & Thompson, J. S. (2003). On specifying the null model for incremental fit indices in structural equation modeling. Psychological Methods, 8(1), 16-37.
}
\examples{

## -- Example A: Measurement Invariance Test Across Groups -- ##

# Data file is "Example.A"

# Specify the measurement model - Model.A
Model.A <- '
       WorkLifeConflict =~ R45a + R45b + R45c + R45d + R45e
       Engagement =~ R90a + R90b + R90c
       Wellbeing =~ R87a + R87b + R87c + R87d + R87e
'

## ===== Full Measurement Invariance Test ===== ##
Full_MEI(Model.A, Example.A, Groups="Region")

}
