% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI.R
\name{LGCompareLoadings}
\alias{LGCompareLoadings}
\title{Metric Invariance for Longitudinal Panel Data}
\usage{
LGCompareLoadings(
  model,
  data.source,
  Cluster = "NULL",
  no.waves = 3,
  Bootstrap = 0,
  Type1 = 0.05,
  Type1Adj = "PFDR",
  BMSC = "SRMR"
)
}
\arguments{
\item{model}{User-specified measurement model.}

\item{data.source}{A data frame containing the observed variables used in the model.}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}

\item{no.waves}{Number of waves for comparisons.}

\item{Bootstrap}{Number of bootstrap samples, must be between 500 and 10,000. If not specified, 1 million Monte Carlo simulated samples (default) will be used.}

\item{Type1}{Overall Type I error rate (default is 0.05) for identifying non-invariant items in the List and Delete method.}

\item{Type1Adj}{Adjustment of Type I error rate for multiple tests for each construct. Default is Possible False Discovery Rate (PFDR = Type I error rate/No. of freely estimated factor loadings/intercepts across all waves), can also use (BON)ferroni adjustment (Type I error rate/No. of pairwise comparisons), or "NULL".}

\item{BMSC}{Best Model Selection Criterion used to select among models with the same number of estimated parameters. "ChiSquare" = minimum Chi-Square value, "CFI" = maximum CFI value, "RMSEA" = minimum RMSEA value, and "SRMR" = minimum SRMR value (default).}
}
\value{
Partial metric invariance model in the PMI.txt file.
}
\description{
Conduct metric invariance test and identify a partial metric invariance model for longitudinal panel data.
}
\details{
Requires defining the measurement model only once. All indicators should have the suffix _T1, _T2 and _T3 (e.g., x1_T1, x1_T2, x1_T3) in the data file to indicate when the indicators were measured.

Residuals of indicators are covaried across time automatically.
}
\examples{

## == Example B - Panel Data in Longitudinal Studies == ##

# Data file is "Example.B"

## Specify the measurement model - Model.B ##
Model.B <- '
  SWLS =~ x1 + x2 + x3 + x4 + x5
'

## ===== Compare Factor Loadings ===== ##
LGCompareLoadings(Model.B, Example.B, no.waves=3, Type1=0.05, Type1Adj="PFDR", BMSC="SRMR")


}
