% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI-package.R
\docType{package}
\name{MEI-package}
\alias{MEI}
\alias{MEI-package}
\title{MEI: Measurement Equivalence/Invariance Test}
\description{
Full measurement equivalence/invariance test
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gche935/MEI}
  \item Report bugs at \url{https://github.com/gche935/MEI/issues}
}

}
\keyword{internal}
