% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI.R
\name{MLCompareLoadings}
\alias{MLCompareLoadings}
\title{Metric Invariance Test Between Level 1 and Level 2}
\usage{
MLCompareLoadings(
  model,
  data.source,
  Cluster = "NULL",
  Type1 = 0.05,
  Type1Adj = "PFDR",
  BMSC = "SRMR"
)
}
\arguments{
\item{model}{User-specified measurement model.}

\item{data.source}{A data frame containing the observed variables used in the model.}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}

\item{Type1}{Overall Type I error rate (default is 0.05) for identifying non-invariant items in the List and Delete method.}

\item{Type1Adj}{Adjustment of Type I error rate for multiple tests for each construct. Default is Possible False Discovery Rate (PFDR = Type I error rate/No. of freely estimated factor loadings/intercepts across all groups), can also use (BON)ferroni adjustment (Type I error rate/No. of pairwise comparisons), or "NULL".}

\item{BMSC}{Best Model Selection Criterion used to select among models with the same number of estimated parameters. "ChiSquare" = minimum Chi-Square value, "CFI" = maximum CFI value, "RMSEA" = minimum RMSEA value, and "SRMR" = minimum sum of SRMR-within and SRMR-between values (default).}
}
\value{
Partial metric invariance model in PMI.txt file.
}
\description{
Conduct configural invariance and full metric invariance test, and identify a partial metric invariance model for multilevel data.
}
\details{
Reference: Measurement Equivalence/Invariance Test based on "Cheung, G. W. & Lau, R. S. (2012).  A direct comparison approach for testing measurement invariance.  Organizational Research Methods, 15, 167-198."

Since all Level 1 (within-group) variables are group-mean centered, intercepts and latent means of all variables at Level 1 are set to zero. Hence, only configural and metric invariance across levels are tested.
Bootstrapping is not available for a multi-level model.

Define the measurement model only once without specifying the level, and the function will compare the model across levels.
}
\examples{

## == Example D - Multilevel Confirmatory Factor Analysis == ##
# Data file is "Example.D"; cluster variable is "ID"

## Specify the measurement model - Model.D ##
Model.D <- 'OLBI =~ x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8'

## ===== Compare Loadings ===== ##
MLCompareLoadings(Model.D, Example.D, Cluster="ID", Type1=0.05, Type1Adj="NULL", BMSC="SRMR")

}
