% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI.R
\name{CompareLoadings}
\alias{CompareLoadings}
\title{Metric Invariance Test}
\usage{
CompareLoadings(
  model,
  data.source,
  Groups,
  Cluster = "NULL",
  Bootstrap = 0,
  Type1 = 0.05,
  Type1Adj = "PFDR",
  BMSC = "SRMR"
)
}
\arguments{
\item{model}{User-specified CFA model.}

\item{data.source}{A data frame containing the observed variables used in the model.}

\item{Groups}{Grouping variable for cross-group comparisons.}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}

\item{Bootstrap}{Number of bootstrap samples, must be between 500 and 10,000. If not specified, 1 million Monte Carlo simulated samples (default) will be used.}

\item{Type1}{Overall Type I error rate (default is 0.05) for identifying non-invariant items in the List and Delete method.}

\item{Type1Adj}{Adjustment of Type I error rate for multiple tests for each construct. Default is Possible False Discovery Rate (PFDR = Type I error rate/No. of freely estimated factor loadings/intercepts across all groups), can also use (BON)ferroni adjustment (Type I error rate/No. of pairwise comparisons), or "NULL".}

\item{BMSC}{Best Model Selection Criterion used to select among models with the same number of estimated parameters. "ChiSquare" = minimum Chi-Square value, "CFI" = maximum CFI value, "RMSEA" = minimum RMSEA value, and "SRMR" = minimum SRMR value (default).}
}
\value{
Partial metric invariance model in the PMI.txt file.
}
\description{
Conduct a full metric invariance test and identify a partial metric invariance model for data from independent groups.
}
\details{
Reference: Measurement Equivalence/Invariance Test based on "Cheung, G. W. & Lau, R. S. (2012).  A direct comparison approach for testing measurement invariance.  Organizational Research Methods, 15, 167-198."

Reference for second-order factor ME/I Test: Cheung, G. W. (2008). Testing Equivalence in the Structure, Means, and Variances of Higher-Order Constructs With Structural Equation Modeling. Organizational Research Methods, 11(3), 593-613.
}
\examples{

## -- Example A: Measurement Invariance Test Across Groups -- ##

# Data file is "Example.A"

# Specify the measurement model - Model.A
Model.A <- '
       WorkLifeConflict =~ R45a + R45b + R45c + R45d + R45e
       Engagement =~ R90a + R90b + R90c
       Wellbeing =~ R87a + R87b + R87c + R87d + R87e
'

## Not run:
## ===== Full Measurement Invariance Test ===== ##
Full_MEI(Model.A, Example.A, Groups="Region")
## End (Not run)

## ===== Compare Loadings ===== ##
CompareLoadings(Model.A, Example.A, Groups="Region", Type1=0.05, Type1Adj="PFDR", BMSC="SRMR")

}
