% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI - Copy.R
\name{CompareLoadings}
\alias{CompareLoadings}
\title{Metric Invariance Test}
\usage{
CompareLoadings(
  model,
  data.source,
  Groups,
  Cluster = "NULL",
  Bootstrap = 0,
  alpha = 0.01
)
}
\arguments{
\item{model}{User-specified CFA model}

\item{data.source}{A data frame containing the observed variables used in the model}

\item{Groups}{Grouping variable for cross-group comparisons}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}

\item{Bootstrap}{Number of bootstrap samples, must be between 500 and 5,000. If not specified, the Monte Carlo simulation (Default) will be used instead of Bootstrapping}

\item{alpha}{Type I error rate for identifying non-invariant items in the List and Delete method. Default is 0.01 (0.05 for MLCFA).  Can also use Bonferroni adjustment (Type I error /No. of comparisons)}
}
\value{
partial metric invariance model in PMI.txt file
}
\description{
Conduct a full metric invariance test and identify a partial metric invariance model
}
\details{
Reference: Measurement Equivalence/Invariance Test based on "Cheung, G. W. & Lau, R. S. (2012).  A direct comparison approach for testing measurement invariance.  Organizational Research Methods, 15, 167-198."
}
\examples{

## -- Example A: Measurement Invariance Test Across Groups -- ##

# Data file is "Example.A"

# Specify the measurement model - Model.A
Model.A <- '
       WorkLifeConflict =~ R45a + R45b + R45c + R45d + R45e
       Engagement =~ R90a + R90b + R90c
       Wellbeing =~ R87a + R87b + R87c + R87d + R87e
'

## ===== Compare Loadings ===== ##
CompareLoadings(Model.A, Example.A, Groups = "Region", alpha = 0.001)

}
