% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI.R
\name{CompareMeans}
\alias{CompareMeans}
\title{Scalar Invariance Test and Compare Latent Means}
\usage{
CompareMeans(
  model.PMI,
  data.source,
  Groups,
  Cluster = "NULL",
  Bootstrap = 0,
  alpha = 0.01
)
}
\arguments{
\item{model.PMI}{Partial metric invariance model (PMI.Model.R from CompareLoadings() or user-specified).}

\item{data.source}{A data frame containing the observed variables used in the model.}

\item{Groups}{Grouping variable for cross-group comparisons.}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}

\item{Bootstrap}{Number of bootstrap samples, must be between 500 and 5,000. If not specified, the Monte Carlo simulation (default) will be used.}

\item{alpha}{Type I error rate (default is 0.01) for identifying non-invariant items in the List and Delete method. Can also use Bonferroni adjustment (Type I error /No. of comparisons).}
}
\value{
Partial scalar invariance model in PSI.txt file and results of latent means comparisons.
}
\description{
Conduct scalar invariance test, identify partial scalar invariance model, and compare latent means.
}
\details{
Reference: Measurement Equivalence/Invariance Test based on "Cheung, G. W. & Lau, R. S. (2012).  A direct comparison approach for testing measurement invariance.  Organizational Research Methods, 15, 167-198."
}
\examples{

## Not run:
## ===== Full Measurement Invariance Test ===== ##
Full_MEI(Model.A, Example.A, Groups = "Region")

## ===== Compare Loadings ===== ##
CompareLoadings(Model.A, Example.A, Groups = "Region", alpha = 0.001)
## End (Not run)

## ===== Compare Intercepts and Latent Means ===== ##
CompareMeans(PMI.Model.R, Example.A, Groups = "Region", alpha = 0.001)

}
