% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI.R
\name{CompareMeans}
\alias{CompareMeans}
\title{Scalar Invariance Test and Compare Latent Means}
\usage{
CompareMeans(
  model.PMI,
  data.source,
  Groups,
  Cluster = "NULL",
  Bootstrap = 0,
  Type1 = 0.05,
  Type1Adj = "PFDR",
  BMSC = "SRMR"
)
}
\arguments{
\item{model.PMI}{Partial metric invariance model (PMI.Model.R from CompareLoadings() or user-specified).}

\item{data.source}{A data frame containing the observed variables used in the model.}

\item{Groups}{Grouping variable for cross-group comparisons.}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}

\item{Bootstrap}{Number of bootstrap samples, must be between 500 and 10,000. If not specified, 1 million Monte Carlo simulated samples (default) will be used.}

\item{Type1}{Overall Type I error rate (default is 0.05) for identifying non-invariant items in the List and Delete method.}

\item{Type1Adj}{Adjustment of Type I error rate for multiple tests for each construct. Default is Possible False Discovery Rate (PFDR = Type I error rate/No. of freely estimated factor loadings/intercepts across all groups), can also use (BON)ferroni adjustment (Type I error rate/No. of pairwise comparisons), or "NULL".}

\item{BMSC}{Best Model Selection Criterion used to select among models with the same number of estimated parameters. "ChiSquare" = minimum Chi-Square value, "CFI" = maximum CFI value, "RMSEA" = minimum RMSEA value, and "SRMR" = minimum SRMR value (default).}
}
\value{
Partial scalar invariance model in PSI.txt file and results of latent means comparisons.
}
\description{
Conduct scalar invariance test, identify partial scalar invariance model, and compare latent means.
}
\details{
Reference: Measurement Equivalence/Invariance Test based on "Cheung, G. W. & Lau, R. S. (2012).  A direct comparison approach for testing measurement invariance.  Organizational Research Methods, 15, 167-198."
}
\examples{

## Not run:
## ===== Full Measurement Invariance Test ===== ##
Full_MEI(Model.A, Example.A, Groups = "Region")

## ===== Compare Loadings ===== ##
CompareLoadings(Model.A, Example.A, Groups = "Region", Type1 = 0.05, Type1Adj = "PFDR", BMSC="SRMR")
## End (Not run)

## ===== Compare Intercepts and Latent Means ===== ##
CompareMeans(PMI.Model.R, Example.A, Groups = "Region", Type1 = 0.05, Type1Adj = "PFDR", BMSC="SRMR")

}
