% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI.R
\name{CompareParameters}
\alias{CompareParameters}
\title{Compare Defined Parameters Across Groups}
\usage{
CompareParameters(
  model.PMI,
  model.PATH,
  data.source,
  Groups,
  Cluster = "NULL",
  Bootstrap = 0
)
}
\arguments{
\item{model.PMI}{Partial metric invariance model (PMI.Model.R from CompareLoadings() or user-specified).}

\item{model.PATH}{model with defined parameters.}

\item{data.source}{A data frame containing the observed variables used in the model.}

\item{Groups}{Grouping variable for cross-group comparisons.}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}

\item{Bootstrap}{Number of bootstrap samples, must be between 500 and 5,000. If not specified, 1 million Monte Carlo simulated samples (default) will be used.}
}
\value{
Estimates and confidence intervals for defined parameters in each group and comparisons of defined parameters across groups.
}
\description{
Compare defined parameters across groups, e.g., direct, indirect and total effects.
}
\details{
Reference: Lau, R. S. & Cheung, G. W. (2012). Estimating and comparing specific mediation effects in complex latent variable models. Organizational Research Methods, 15, 3-16.
}
\examples{

## Specify the measurement model - Model.A
Model.A <- '
       WorkLifeConflict =~ R45a + R45b + R45c + R45d + R45e
       Engagement =~ R90a + R90b + R90c
       Wellbeing =~ R87a + R87b + R87c + R87d + R87e
'

## Not run:
## ===== Full Measurement Invariance Test ===== ##
Full_MEI(Model.A, Example.A, Groups = "Region")

## ===== Compare Loadings ===== ##
CompareLoadings(Model.A, Example.A, Groups = "Region", Type1 = 0.05, Type1Adj = "PFDR")
## End (Not run)

## ===== Compare Parameters ===== ##
# -- Specify Path model - model.PATH (model.DP) [OrgSize and Tenure are control variables] -- #
model.DP <- '
  Wellbeing ~ Xb1*Engagement + Xc1*WorkLifeConflict + Xd1*OrgSize + Xe1*Tenure
  Engagement ~ Xa1*WorkLifeConflict

 # Defined Parameters #
 IndirectP := Xa1*Xb1  # Indirect effect
 DirectP := Xc1  # Direct effect
 Total := Xa1*Xb1 + Xc1  # Total effect
'

# -- Run function CompareParameters using  Monte Carlo simulation -- #
CompareParameters(PMI.Model.R, model.DP, Example.A, Groups = "Region")

}
\keyword{internal}
