% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI.R
\name{Full_MEI}
\alias{Full_MEI}
\title{Full Measurement Invariance Test}
\usage{
Full_MEI(model, data.source, Groups, Cluster = "NULL")
}
\arguments{
\item{model}{User-specified CFA model}

\item{data.source}{A data frame containing the observed variables used in the model}

\item{Groups}{Grouping variable for cross-group comparisons}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}
}
\value{
lavaan outputs, model fit for configural, metric and scalar invariance models, summary of fit statistics for each model
}
\description{
Conduct configural invariance, full metric invariance and full scalar invariance tests
}
\details{
Missing values are handled with full-information maximum likelihood.
}
\examples{

## -- Example A: Measurement Invariance Test Across Groups -- ##

# Data file is "Example.A"

# Specify the measurement model - Model.A
Model.A <- '
       WorkLifeConflict =~ R45a + R45b + R45c + R45d + R45e
       Engagement =~ R90a + R90b + R90c
       Wellbeing =~ R87a + R87b + R87c + R87d + R87e
'

## ===== Full Measurement Invariance Test ===== ##
Full_MEI(Model.A, Example.A, Groups = "Region")

}
