% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI.R
\name{LGCompareLoadings}
\alias{LGCompareLoadings}
\title{Metric Invariance Test Across Time or Sources}
\usage{
LGCompareLoadings(
  model,
  data.source,
  Cluster = "NULL",
  no.waves = 3,
  alpha = 0.01
)
}
\arguments{
\item{model}{User-specified CFA model}

\item{data.source}{A data frame containing the observed variables used in the model}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}

\item{no.waves}{Number of waves for comparisons}

\item{alpha}{Type I error rate for identifying non-invariant items in the List and Delete method. Default is 0.01 (0.05 for MLCFA).  Can also use Bonferroni adjustment (Type I error /No. of comparisons)}

\item{Bootstrap}{Number of bootstrap samples, must be between 500 and 5,000. If not specified, the Monte Carlo simulation (Default) will be used instead of Bootstrapping}
}
\value{
estimates and confidence intervals for defined parameters in each group and comparisons of defined parameters across groups
}
\description{
Conduct metric invariance test and identify non-invariant items across time or different sources.
}
\details{
Requires defining the measurement model only once. All indicators should have the suffix _T1, _T2 and _T3 (e.g., x1_T1, x1_T2, x1_T3) in the data file to indicate when the item was measured.

Residuals of indicators are covaried across time automatically
}
\examples{

## == Example B - Panel Data in Longitudinal Studies == ##

# Data file is "Example.B"

## Specify the measurement model - Model.B ##
Model.B <- '
  SWLS =~ x1 + x2 + x3 + x4 + x5
'

## ===== Compare Factor Loadings ===== ##
LGCompareLoadings(Model.B, Example.B, no.waves = 3, alpha =  0.01)


## == Example C - Non-independent Data from two sources == ##

# Data file is "Example.C"

## Specify the measurement model - Model.C ##
Model.C <- '
     External =~ x1 + x2 + x3 + x4 + x5 + x6
     Internal =~ x7 + x8 + x9 + x10
'

## ===== Compare Factor Loadings ===== ##
LGCompareLoadings(Model.C, Example.C, no.waves = 2, alpha = 0.01)

}
