% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI.R
\name{MLCompareLoadings}
\alias{MLCompareLoadings}
\title{Metric Invariance Test Between Level 1 and Level 2}
\usage{
MLCompareLoadings(model, data.source, Cluster = "NULL", alpha = 0.05)
}
\arguments{
\item{model}{User-specified CFA model}

\item{data.source}{A data frame containing the observed variables used in the model}

\item{Cluster}{Cluster variable for nested data. The Monte Carlo simulation method should be used for nested data.}

\item{alpha}{Type I error rate for identifying non-invariant items in the List and Delete method. Default is 0.01 (0.05 for MLCFA).  Can also use Bonferroni adjustment (Type I error /No. of comparisons)}
}
\value{
partial metric invariance model in PMI.txt file
}
\description{
Conduct configural invariance and full metric invariance test, and identify partial metric invariance model
}
\details{
Reference: Measurement Equivalence/Invariance Test based on "Cheung, G. W. & Lau, R. S. (2012).  A direct comparison approach for testing measurement invariance.  Organizational Research Methods, 15, 167-198."

Since all Level 1 (within-group) variables are group-mean centered, intercepts and latent means of all variables at Level 1 are set to zero. Hence, only configural and metric invariance across levels are tested
Bootstrapping is not available for multi-level model

Define the measurement model only once without specifying the level, and the function will compare the model across levels
}
\examples{

## == Example D - Multilevel Confirmatory Factor Analysis == ##
# Data file is "Example.A"; cluster variable is "ID"

## Specify the measurement model - Model.D ##
Model.D <- 'OLBI =~ x1 + x2 + x3 + x4 + x5 + x6 + x7 + x8'

## ===== Compare Loadings ===== ##
MLCompareLoadings(Model.D, Example.D, Cluster = "ID", alpha = 0.05)

}
